monitoring/autoheal/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
prometheus-client==0.20.0
requests==2.31.0
python-jose[cryptography]==3.3.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-requests==0.42b0
